add save history function to model

+++ jabble/model.py
@@ -227,8 +227,8 @@ class ContainerModel(Model):
 
     def save_history(self,p):
         for i,model in enumerate(self.models):
-            model.save_history(p[np.arange(np.sum(self.parameters_per_model[:i]), \
-                                            np.sum(self.parameters_per_model[:i+1]),dtype=int)])
+            model.save_history(p[np.arange(np.sum(self.parameters_per_model[:i]).astype(int), \
+                                            np.sum(self.parameters_per_model[:i+1]).astype(int),dtype=int)])
 
 
 class CompositeModel(ContainerModel):
